Dollar-exchange-rate/server/server_requirements.txt
fastapi
uvicorn
python-dateutil
python-dotenv
requests
python-dotenv
python-dateutil
